Refactor: Restructure project for modular Manim development
This commit introduces a more organized and modular project structure
to facilitate the creation of complex Manim animations.

Key changes include:

1.  **Directory Reorganization:**
    *   I created a `scenes/` directory for all your Manim scene scripts.
    *   I created a `src/` directory for your custom Python modules, utility functions (e.g., `manim_utils.py`, `structures.py`), and other reusable code. I also added an `__init__.py` file to `src/` to make it a Python package.
    *   I created an `assets/` directory for static files like images, sound files, etc.

2.  **File Migration:**
    *   I moved existing Python scripts from the root directory into the appropriate new directories (`scenes/` or `src/`).

3.  **Configuration File Updates:**
    *   `.gitignore`: I enhanced this to include standard Python, Manim, and OS-specific exclusions (e.g., `__pycache__/`, `output/`, `venv/`).
    *   `README.md`: I significantly updated this to reflect the new project structure, with revised instructions on how to set up the environment, where to place new files, and how to run example scenes with their new paths.
    *   `requirements.txt`: I reviewed this and it seems adequate for now.

4.  **Import Adjustments:**
    *   I scanned your scene files and updated import statements to correctly reference modules that were moved to the `src/` directory (e.g., `import manim_utils` changed to `from src import manim_utils` or `from src.manim_utils import ...`). This affected four files.

5.  **Backup:** I backed up your original `README.md`, `requirements.txt`, and `.gitignore` files with a `.bak` extension.

This restructuring aims to align with common practices in Python and Manim project organization, making your codebase more maintainable and scalable. I did not complete the "Test a Few Scenes" step in this session.
