+++ .gitignore.bak
@@ -1,0 +1,5 @@
+/media/
+*.mp4
+*.mp3
+*.pyc
+/cards/
