+++ requirements.txt.bak
@@ -1,0 +1,4 @@
+manim
+numpy
+scipy
+pyglet
