+++ scenes/draw.py
