+++ scenes/sweeping.py
